+++ webview-dist/index.d.ts
@@ -1,30 +1,30 @@
-export declare function all_sys_info(): Promise<unknown>;
-export declare function total_memory(): Promise<unknown>;
-export declare function used_memory(): Promise<unknown>;
-export declare function total_swap(): Promise<unknown>;
-export declare function used_swap(): Promise<unknown>;
-export declare function memory_info(): Promise<unknown>;
+export declare function allSysInfo(): Promise<unknown>;
+export declare function totalMemory(): Promise<unknown>;
+export declare function usedMemory(): Promise<unknown>;
+export declare function totalSwap(): Promise<unknown>;
+export declare function usedSwap(): Promise<unknown>;
+export declare function memoryInfo(): Promise<unknown>;
 export declare function hostname(): Promise<unknown>;
 export declare function name(): Promise<unknown>;
-export declare function kernel_version(): Promise<unknown>;
-export declare function os_version(): Promise<unknown>;
-export declare function static_info(): Promise<unknown>;
+export declare function kernelVersion(): Promise<unknown>;
+export declare function osVersion(): Promise<unknown>;
+export declare function staticInfo(): Promise<unknown>;
 export declare function components(): Promise<unknown>;
 export declare function cpus(): Promise<unknown>;
-export declare function cpu_count(): Promise<unknown>;
-export declare function cpu_info(): Promise<unknown>;
+export declare function cpuCount(): Promise<unknown>;
+export declare function cpuInfo(): Promise<unknown>;
 export declare function disks(): Promise<unknown>;
 export declare function networks(): Promise<unknown>;
 export declare function processes(): Promise<unknown>;
-export declare function refresh_all(): Promise<unknown>;
-export declare function refresh_memory(): Promise<unknown>;
-export declare function refresh_cpu(): Promise<unknown>;
-export declare function refresh_system(): Promise<unknown>;
-export declare function refresh_components(): Promise<unknown>;
-export declare function refresh_components_list(): Promise<unknown>;
-export declare function refresh_disks(): Promise<unknown>;
-export declare function refresh_disks_list(): Promise<unknown>;
-export declare function refresh_users_list(): Promise<unknown>;
-export declare function refresh_networks(): Promise<unknown>;
-export declare function refresh_networks_list(): Promise<unknown>;
-export declare function refresh_processes(): Promise<unknown>;
+export declare function refreshAll(): Promise<unknown>;
+export declare function refreshMemory(): Promise<unknown>;
+export declare function refreshCpu(): Promise<unknown>;
+export declare function refreshSystem(): Promise<unknown>;
+export declare function refreshComponents(): Promise<unknown>;
+export declare function refreshComponentsList(): Promise<unknown>;
+export declare function refreshDisks(): Promise<unknown>;
+export declare function refreshDisksList(): Promise<unknown>;
+export declare function refreshUsersList(): Promise<unknown>;
+export declare function refreshNetworks(): Promise<unknown>;
+export declare function refreshNetworksList(): Promise<unknown>;
+export declare function refreshProcesses(): Promise<unknown>;
