+++ webview-dist/index.js
@@ -1,1 +1,1 @@
-var n=Object.defineProperty;function e(n,e=!1){let t=window.crypto.getRandomValues(new Uint32Array(1))[0],r=`_${t}`;return Object.defineProperty(window,r,{value:t=>(e&&Reflect.deleteProperty(window,r),n?.(t)),writable:!1,configurable:!0}),t}async function t(n,t={}){return new Promise(((r,i)=>{let s=e((n=>{r(n),Reflect.deleteProperty(window,`_${o}`)}),!0),o=e((n=>{i(n),Reflect.deleteProperty(window,`_${s}`)}),!0);window.__TAURI_IPC__({cmd:n,callback:s,error:o,...t})}))}function r(n,e="asset"){let t=encodeURIComponent(n);return navigator.userAgent.includes("Windows")?`https://${e}.localhost/${t}`:`${e}://localhost/${t}`}function i(){return t("plugin:system-info|all_sys_info")}function s(){return t("plugin:system-info|total_memory")}function o(){return t("plugin:system-info|used_memory")}function u(){return t("plugin:system-info|total_swap")}function f(){return t("plugin:system-info|used_swap")}function l(){return t("plugin:system-info|memory_info")}function c(){return t("plugin:system-info|hostname")}function m(){return t("plugin:system-info|name")}function p(){return t("plugin:system-info|kernel_version")}function y(){return t("plugin:system-info|os_version")}function _(){return t("plugin:system-info|static_info")}function g(){return t("plugin:system-info|components")}function a(){return t("plugin:system-info|cpus")}function d(){return t("plugin:system-info|cpu_count")}function w(){return t("plugin:system-info|cpu_info")}function h(){return t("plugin:system-info|disks")}function k(){return t("plugin:system-info|networks")}function v(){return t("plugin:system-info|processes")}function b(){return t("plugin:system-info|refresh_all")}function P(){return t("plugin:system-info|refresh_memory")}function $(){return t("plugin:system-info|refresh_cpu")}function R(){return t("plugin:system-info|refresh_system")}function A(){return t("plugin:system-info|refresh_components")}function C(){return t("plugin:system-info|refresh_components_list")}function I(){return t("plugin:system-info|refresh_disks")}function U(){return t("plugin:system-info|refresh_disks_list")}function j(){return t("plugin:system-info|refresh_users_list")}function O(){return t("plugin:system-info|refresh_networks")}function x(){return t("plugin:system-info|refresh_networks_list")}function F(){return t("plugin:system-info|refresh_processes")}((e,t)=>{for(var r in t)n(e,r,{get:t[r],enumerable:!0})})({},{convertFileSrc:()=>r,invoke:()=>t,transformCallback:()=>e});export{i as all_sys_info,g as components,d as cpu_count,w as cpu_info,a as cpus,h as disks,c as hostname,p as kernel_version,l as memory_info,m as name,k as networks,y as os_version,v as processes,b as refresh_all,A as refresh_components,C as refresh_components_list,$ as refresh_cpu,I as refresh_disks,U as refresh_disks_list,P as refresh_memory,O as refresh_networks,x as refresh_networks_list,F as refresh_processes,R as refresh_system,j as refresh_users_list,_ as static_info,s as total_memory,u as total_swap,o as used_memory,f as used_swap};
+var n=Object.defineProperty;function e(n,e=!1){let t=window.crypto.getRandomValues(new Uint32Array(1))[0],r=`_${t}`;return Object.defineProperty(window,r,{value:t=>(e&&Reflect.deleteProperty(window,r),n?.(t)),writable:!1,configurable:!0}),t}async function t(n,t={}){return new Promise(((r,i)=>{let s=e((n=>{r(n),Reflect.deleteProperty(window,`_${o}`)}),!0),o=e((n=>{i(n),Reflect.deleteProperty(window,`_${s}`)}),!0);window.__TAURI_IPC__({cmd:n,callback:s,error:o,...t})}))}function r(n,e="asset"){let t=encodeURIComponent(n);return navigator.userAgent.includes("Windows")?`https://${e}.localhost/${t}`:`${e}://localhost/${t}`}function i(){return t("plugin:system-info|all_sys_info")}function s(){return t("plugin:system-info|total_memory")}function o(){return t("plugin:system-info|used_memory")}function u(){return t("plugin:system-info|total_swap")}function f(){return t("plugin:system-info|used_swap")}function l(){return t("plugin:system-info|memory_info")}function c(){return t("plugin:system-info|hostname")}function m(){return t("plugin:system-info|name")}function p(){return t("plugin:system-info|kernel_version")}function y(){return t("plugin:system-info|os_version")}function _(){return t("plugin:system-info|static_info")}function g(){return t("plugin:system-info|components")}function a(){return t("plugin:system-info|cpus")}function d(){return t("plugin:system-info|cpu_count")}function w(){return t("plugin:system-info|cpu_info")}function h(){return t("plugin:system-info|disks")}function k(){return t("plugin:system-info|networks")}function v(){return t("plugin:system-info|processes")}function b(){return t("plugin:system-info|refresh_all")}function P(){return t("plugin:system-info|refresh_memory")}function $(){return t("plugin:system-info|refresh_cpu")}function R(){return t("plugin:system-info|refresh_system")}function A(){return t("plugin:system-info|refresh_components")}function C(){return t("plugin:system-info|refresh_components_list")}function I(){return t("plugin:system-info|refresh_disks")}function U(){return t("plugin:system-info|refresh_disks_list")}function j(){return t("plugin:system-info|refresh_users_list")}function O(){return t("plugin:system-info|refresh_networks")}function x(){return t("plugin:system-info|refresh_networks_list")}function F(){return t("plugin:system-info|refresh_processes")}((e,t)=>{for(var r in t)n(e,r,{get:t[r],enumerable:!0})})({},{convertFileSrc:()=>r,invoke:()=>t,transformCallback:()=>e});export{i as allSysInfo,g as components,d as cpuCount,w as cpuInfo,a as cpus,h as disks,c as hostname,p as kernelVersion,l as memoryInfo,m as name,k as networks,y as osVersion,v as processes,b as refreshAll,A as refreshComponents,C as refreshComponentsList,$ as refreshCpu,I as refreshDisks,U as refreshDisksList,P as refreshMemory,O as refreshNetworks,x as refreshNetworksList,F as refreshProcesses,R as refreshSystem,j as refreshUsersList,_ as staticInfo,s as totalMemory,u as totalSwap,o as usedMemory,f as usedSwap};
