+++ webview-src/index.ts
@@ -1,22 +1,22 @@
 import { invoke } from "@tauri-apps/api/tauri";
 
-export function all_sys_info() {
+export function allSysInfo() {
   return invoke("plugin:system-info|all_sys_info");
 }
 // Memory
-export function total_memory() {
+export function totalMemory() {
   return invoke("plugin:system-info|total_memory");
 }
-export function used_memory() {
+export function usedMemory() {
   return invoke("plugin:system-info|used_memory");
 }
-export function total_swap() {
+export function totalSwap() {
   return invoke("plugin:system-info|total_swap");
 }
-export function used_swap() {
+export function usedSwap() {
   return invoke("plugin:system-info|used_swap");
 }
-export function memory_info() {
+export function memoryInfo() {
   return invoke("plugin:system-info|memory_info");
 }
 // Static Info
@@ -26,13 +26,13 @@ export function hostname() {
 export function name() {
   return invoke("plugin:system-info|name");
 }
-export function kernel_version() {
+export function kernelVersion() {
   return invoke("plugin:system-info|kernel_version");
 }
-export function os_version() {
+export function osVersion() {
   return invoke("plugin:system-info|os_version");
 }
-export function static_info() {
+export function staticInfo() {
   return invoke("plugin:system-info|static_info");
 }
 // Components
@@ -43,10 +43,10 @@ export function components() {
 export function cpus() {
   return invoke("plugin:system-info|cpus");
 }
-export function cpu_count() {
+export function cpuCount() {
   return invoke("plugin:system-info|cpu_count");
 }
-export function cpu_info() {
+export function cpuInfo() {
   return invoke("plugin:system-info|cpu_info");
 }
 // Disk
@@ -62,39 +62,39 @@ export function processes() {
   return invoke("plugin:system-info|processes");
 }
 // Refresh
-export function refresh_all() {
+export function refreshAll() {
   return invoke("plugin:system-info|refresh_all");
 }
-export function refresh_memory() {
+export function refreshMemory() {
   return invoke("plugin:system-info|refresh_memory");
 }
-export function refresh_cpu() {
+export function refreshCpu() {
   return invoke("plugin:system-info|refresh_cpu");
 }
-export function refresh_system() {
+export function refreshSystem() {
   return invoke("plugin:system-info|refresh_system");
 }
-export function refresh_components() {
+export function refreshComponents() {
   return invoke("plugin:system-info|refresh_components");
 }
-export function refresh_components_list() {
+export function refreshComponentsList() {
   return invoke("plugin:system-info|refresh_components_list");
 }
-export function refresh_disks() {
+export function refreshDisks() {
   return invoke("plugin:system-info|refresh_disks");
 }
-export function refresh_disks_list() {
+export function refreshDisksList() {
   return invoke("plugin:system-info|refresh_disks_list");
 }
-export function refresh_users_list() {
+export function refreshUsersList() {
   return invoke("plugin:system-info|refresh_users_list");
 }
-export function refresh_networks() {
+export function refreshNetworks() {
   return invoke("plugin:system-info|refresh_networks");
 }
-export function refresh_networks_list() {
+export function refreshNetworksList() {
   return invoke("plugin:system-info|refresh_networks_list");
 }
-export function refresh_processes() {
+export function refreshProcesses() {
   return invoke("plugin:system-info|refresh_processes");
 }
